+++ website_support/models/res_partner.py
@@ -9,6 +9,7 @@ class ResPartnerTicket(models.Model):
     support_ticket_count = fields.Integer(compute="_count_support_tickets", string="Ticket Count")
     new_support_ticket_count = fields.Integer(compute="_count_new_support_tickets", string="New Ticket Count")
     support_ticket_string = fields.Char(compute="_compute_support_ticket_string", string="Support Ticket String")
+    stp_ids = fields.Many2many('res.partner', 'stp_res_partner_rel', 'stp_p1_id', 'stp_p2_id', string="Support Ticket Access Accounts", help="Can view support tickets from other contacts")
     
     @api.one
     @api.depends('support_ticket_ids')